Add unique_id method to User
This give access to the 'sub' claim from the OIDC Issuer, as, per
specification, that's the only one that will not change (see 5.7 of
https://openid.net/specs/openid-connect-core-1_0-final.html).

Some downstream application will have to be fixed to use this
instead of the nickname property as a compliant OIDC provider
can change "nickname" or "email".

The "sub" claim is always returned in UserInfo response (5.3.2 of the
spec).